+++ src/main/java/com/appli/springjwt/repository/MatiereconcourstciRepository.java
@@ -4,6 +4,7 @@ import com.appli.springjwt.models.Matiereconcourstci;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -14,8 +15,8 @@ public interface MatiereconcourstciRepository extends JpaRepository<Matiereconco
     Optional<Matiereconcourstci> findByNomMCTCI(String nomMCTCI);
 
     Boolean existsByNomMCTCI(String nomMCTCI);
-    Boolean existsByNomMCTCIAndCreditMCTCI(String nomMCTCI, Byte creditMCTCI);
-    Optional<Matiereconcourstci> findByNomMCTCIAndCreditMCTCI(String nomMCTCI, Byte creditMCTCI);
+    Boolean existsByNomMCTCIAndCreditMCTCI(String nomMCTCI, BigDecimal creditMCTCI);
+    Optional<Matiereconcourstci> findByNomMCTCIAndCreditMCTCI(String nomMCTCI, BigDecimal creditMCTCI);
 
     ArrayList<Matiereconcourstci> findAllById(Integer id);
